# perf: micro-optimize core.encode 