--- a/core/bindings.rs
+++ b/core/bindings.rs
@@ -864,9 +864,12 @@ fn encode(
     }
   };
   let text_str = text.to_rust_string_lossy(scope);
-  let zbuf: ZeroCopyBuf = text_str.into_bytes().into();
-
-  rv.set(to_v8(scope, zbuf).unwrap())
+  let bytes: Box<[u8]> = text_str.into_bytes().into_boxed_slice();
+  let len = bytes.len();
+  let bs = v8::ArrayBuffer::new_backing_store_from_boxed_slice(bytes).make_shared();
+  let buffer = v8::ArrayBuffer::with_backing_store(scope, &bs);
+  let u8array = v8::Uint8Array::new(scope, buffer, 0, len).unwrap();
+  rv.set(u8array.into())
 }
 
 fn decode(