`main`:
```
test bench_utf8_decode_12_b  ... bench:         187 ns/iter (+/- 12)
test bench_utf8_decode_12_kb ... bench:       4,589 ns/iter (+/- 676)
test bench_utf8_decode_12_mb ... bench:   1,502,699 ns/iter (+/- 28,120)
test bench_utf8_encode_12_b  ... bench:         450 ns/iter (+/- 42)
test bench_utf8_encode_12_kb ... bench:       5,586 ns/iter (+/- 295)
test bench_utf8_encode_12_mb ... bench:   4,767,425 ns/iter (+/- 9,069)
```

This PR:
```
test bench_utf8_decode_12_b  ... bench:         176 ns/iter (+/- 3)
test bench_utf8_decode_12_kb ... bench:       3,268 ns/iter (+/- 80)
test bench_utf8_decode_12_mb ... bench:   1,441,622 ns/iter (+/- 19,107)
test bench_utf8_encode_12_b  ... bench:         354 ns/iter (+/- 15)
test bench_utf8_encode_12_kb ... bench:       5,401 ns/iter (+/- 109)
test bench_utf8_encode_12_mb ... bench:   4,760,358 ns/iter (+/- 26,787)
```