@aapoalas 2022-03-25
LGTM. I'd maybe rename the `bs` into `backing_store` just to be explicit but that's just the nittiest of picks :)